networkDesign: rename `SimulationAlgorithm` to `TransitNetworkDesignAlgorithm`
part of #1559

The current `SimulationAlgorithm` type only applies to network design
for mode transit, so are renamed to `TransitNetworkDesignAlgorithm`.
Files and types are renamed accordingly, though not the descriptor, as
they are far more generic than just algorithm descriptors. They are
actually used also for simulation method options.
