+++ packages/transition-common/src/services/networkDesign/transit/algorithm/EvolutionaryAlgorithm.ts
@@ -5,7 +5,7 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 
-import { SimulationAlgorithmDescriptor } from '../../../simulation/SimulationAlgorithm';
+import { SimulationAlgorithmDescriptor } from '../TransitNetworkDesignAlgorithm';
 
 // Define evolutionary algorithm options
 export type EvolutionaryAlgorithmOptions = {
