+++ packages/transition-common/src/services/networkDesign/transit/algorithm/index.ts
@@ -5,7 +5,7 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 
-import { SimulationAlgorithmDescriptor } from '../../../simulation/SimulationAlgorithm';
+import { SimulationAlgorithmDescriptor } from '../TransitNetworkDesignAlgorithm';
 import { EvolutionaryAlgorithmDescriptor, EvolutionaryAlgorithmOptions } from './EvolutionaryAlgorithm';
 
 /**
