+++ packages/transition-common/src/services/networkDesign/transit/simulationMethod/OdTripSimulationMethod.ts
@@ -5,7 +5,7 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 
-import { SimulationAlgorithmDescriptor } from '../../../simulation/SimulationAlgorithm';
+import { SimulationAlgorithmDescriptor } from '../TransitNetworkDesignAlgorithm';
 
 // Define OD trip simulation options
 export type OdTripSimulationOptions = {
